requirements.txt
celery==5.2.7
cryptography==41.0.0
django-allauth==0.54.0
django-celery==3.1.17
Django==4.0.10
oauthlib==3.2.2
pycparser==2.21
pyparsing==3.0.9
redis==4.5.5
requests-oauthlib==1.3.1
requests==2.31.0
PyMySQL==1.0.3
django-bootstrap-datepicker-plus==5.0.3